Wrap git-show instead of git-log
This only makes a differenvce when run with a single commit argument
(as opposed to a commit range): git-log will walk the history from that
commit and also show its ancestors, whereas git-show will show only that
commit. IMHO, it makes more sense for git-review to follow git-show's
behavior here.