Support for Review-Section syntax
This makes it easy for a commit author to create sections like so:

Review-Section: src/git-review

Extracts the new trailer (interloper?) and aggregates any arguments
(comma- or space-separated) into the --order field

Review-Section: test

An example with the new syntax. This uses exactly the same pattern as
the other commits, so there's not really much to see in the output here.